2020 Day 2
Puzzle:

pt1: Lowest and highest number of times a given letter must appear for the password to be valid.
How many passwords are valid according to their policies?

pt2: Positions in the password, where 1 means the first character, 2 means the second character. Only one must match.
How many passwords are valid according to the new interpretation of the policies?